feat: move bot to game state
